requirements.txt
flask 
flask_sqlalchemy
os
werkzeug.utils
gunicorn
